+++ R/jam-igraph-plot.R
@@ -5,22 +5,134 @@
 #'
 #' Jam wrapper to plot igraph
 #'
-#' This function is a lightweight wrapper around `igraph::plot.igraph()`
-#' intended to handle `rescale=FALSE` properly, which is not done
-#' in the former function (as of January 2020). The desired outcome
-#' is for the `xlim` and `ylim` defaults to be scaled according to the
-#' `igraph` layout. Similarly, the `vertex.size` and `vertex.label.dist`
-#' parameters should also be scaled proportionally.
+#' This function is a drop-in replacement of `igraph::plot.igraph()`,
+#' intended to provide substantially faster vectorized plotting,
+#' to render bundled edges when requested, and
+#' to handle `rescale=FALSE` without requiring further adjustments.
 #'
-#' You can use argument `label_factor_l=list(nodeType=c(Gene=0.01, Set=1))`
-#' to hide labels for Gene nodes, and display labels for Set nodes.
-#' Note that due to a quirk in `igraph`, setting `label.cex=0`
-#' will revert the font to default size, and will not hide the
-#' label.
+#' It also provides some convenient methods to adjust node
+#' size, label font size, and label distance from node center,
+#' based upon node attributes.
+#'
+#' ## vectorized plots
+#'
+#' This function calls `jam_plot_igraph()` as a replacement for
+#' `igraph::plot.igraph()`, and that function implements vectorized
+#' plot features when `vectorized_node_shapes=TRUE` by default:
+#'
+#' 1. When there are multiple different vertex `"shape"` attributes,
+#' this function renders nodes vectorized, one shape at a time.
+#' In this scenario, the original `igraph::plot.igraph()` draws
+#' each individual vertex,
+#' which is substantially slower (minutes compared to 1-2 seconds)
+#' for large `igraph` objects.
+#' 2. When there are multiple font families, labels are rendered in
+#' groups by font family, in order to comply with limitations
+#' in `graphics::text()`. This situation is fairly rare, however
+#' the speed improvement is substantial, again roughly minutes down
+#' to 1-2 seconds. The `igraph::plot.igraph()` renders each node label
+#' individually when there are multiple font families.
+#'
+#' ## rescale=FALSE
+#'
+#' The default `igraph::plot.igraph()` uses `rescale=TRUE`, which
+#' can distort the layout coordinates to fit within a fixed
+#' x- and y-axis range `c(-1, 1)`. When using `rescale=FALSE` the
+#' `xlim` and `ylim` values are not adjusted to the actual coordinate
+#' range. The desired effect from this function `jam_igraph()` is
+#' to apply `aspect=1` (`asp=1`) which fixes the aspect ratio so
+#' the coordinates represent visual Euclidean distance between nodes,
+#' and to define `xlim` and `ylim` to accomodate the full layout.
+#' This function also adjusts node `vertex.size` and `vertex.label.dist`
+#' proportionally.
+#'
+#' ## edge bundling
+#'
+#' When `edge_bundling` is something other than `edge_bundling="none"`,
+#' edge connections between nodes are rendered using a specific
+#' function by drawing curved splines for each bundle of edges.
+#' The approach in `igraph::plot.igraph()` only draws
+#' straight edges between nodes. The recommended method is
+#' `edge_bundling="connections"` which will bundle edges among nodes
+#' that share identical connections.
+#'
+#' ## Adjust node size, label size, label distance
+#'
+#' The following arguments apply scaling to all nodes or edges:
+#'
+#' * `node_factor` - `numeric` multiplied by node size
+#' * `edge_factor` - `numeric` multiplied by edge size
+#' * `label_factor` - `numeric` multiplied by label font size
+#' * `label_dist_factor` - `numeric` multiplied by label distance from
+#' node center
+#'
+#' The following arguments apply scale factor based upon node attribute:
+#'
+#' * `node_factor_l` - `list` of named vectors applied to node size
+#' * `label_factor_l` - `list` of named vectors applied to label font size
+#' * `label_dist_factor_l` - `list` of named vectors applied to label
+#' distance from node center
+#'
+#' The `factor_l` technique is as follows:
+#'
+#' ```R
+#' node_factor_l = list(node_attr_name = c(
+#'    attr_value1 = factor1,
+#'    attr_value2 = factor2))
+#' ```
+#'
+#' A specific example:
+#' ```R
+#' node_factor_l = list(nodeType = c(
+#'    Gene=1.5,
+#'    Set=2))
+#' ```
+#'
+#' In this case, nodes with attribute
+#' `V(g)$nodeType == "Gene"` will use factor `1.5`
+#' Nodes with attribute
+#' `V(g)$nodeType == "Set"` will use factor `2`
+#' All other nodes will not be adjusted.
+#'
+#' ## Other features
+#'
+#' When `use_shadowText=TRUE` node labels call `jamba::shadowText()`
+#' which draws a small partly transparent outline around labels, making
+#' them more legible when they overlap colored nodes. This step
+#' effectively draws each label `n` times, which can slightly slow
+#' the rendering of the overall figure.
+#'
+#' When `pie_to_jampie=TRUE`, any nodes with `shape="pie"` are
+#' changed to `shape="jampie"` for the purpose of rendering pie
+#' shapes in vectorized fashion, instead of being drawn for each
+#' node separately. This change is a substantial improvement in
+#' rendering time.
+#'
+#' Default colors for marked node groups `mark.col` and `mark.border`
+#' when not defined upfront, will call `colorjam::rainbowJam()`
+#' and not `grDevices::rainbow(). The `colorjam::rainbowJam()`
+#' produces more visually distinct categorical colors.
+#' This behavior can be controlled by supplying a `character`
+#' vector with specific colors for `mark.col` and `mark.border`. Note
+#' that the border should match the colors, or it can be set to `"grey45"`
+#' for a generally visible border.
+#'
+#' Optional argument `nodegroups` can be supplied, which is a `list`
+#' of vectors, where each vector represents a group of nodes. The
+#' `nodegroups` can be used with `edge_bundling="nodegroups"` to
+#' define custom edge bundling.
+#'
+#' Finally, individual plot components can be individually disabled:
+#'
+#' * `render_nodes=FALSE`
+#' * `render_edges=FALSE`
+#' * `render_groups=FALSE`
+#' * `render_nodelabels=FALSE`
 #'
 #' @family jam igraph functions
 #' @family jam plot functions
 #'
+#' @inheritParams jam_plot_igraph
 #' @param x `igraph` object to be plotted
 #' @param ... additional arguments are passed to `igraph::plot.igraph()`
 #' @param xlim,ylim default x and y axis limits